requirements.txt
Flask-sqlalchemy
Flask
python-dotenv
flask_cors
Flask-JWT-Extended
itsdangerous
Flask-Mail
flask_limiter
redis